requirements.txt
sqlalchemy>=2.0.0
pandas>=2.0.0
pyyaml>=6.0

pyspark>=3.5.0

snowflake-connector-python>=3.10.0
snowflake-sqlalchemy>=1.6.0

streamlit>=1.40.0
plotly>=5.20.0

